fix: support macOS Catalina
The output of `security find-certificate -apZ` on macOS Catalina prints both SHA-1 and SHA-256 hashes. The current parser does not expect this and fails. Add an optional line in the parser to get the SHA-256 hash if there's such. This way the parser can work with older macOS and with Catalina as well. Add unit test for the change.